requirements.txt
certifi==2021.5.30
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Cors==3.0.10
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
is-disposable-email==1.0.0
itsdangerous==2.0.1
Jinja2==3.0.1
libsass==0.21.0
Markdown==3.3.4
MarkupSafe==2.0.1
Pygments==2.10.0
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
WTForms==2.3.3
wtforms-validators==1.0.0
